# api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LunarBase API",
    "description": "A powerful backend-as-a-service API built with Rust, Axum, and Diesel ORM",
    "contact": {
      "name": "LunarBase Team",
      "email": "contact@lunarbase.dev"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    },
    {
      "url": "https://api.lunarbase.dev",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/collections/{name}": {
      "get": {
        "tags": [
          "Collections"
        ],
        "summary": "Get collection by name",
        "operationId": "get_collection",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Collection name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Collection retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CollectionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Collection not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/permissions/collections/{collection_name}": {
      "get": {
        "tags": [
          "Permissions"
        ],
        "summary": "Get collection permissions",
        "operationId": "get_collection_permissions",
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "description": "Collection name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Collection permissions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Value"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions - Admin only",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Collection not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Permissions"
        ],
        "summary": "Set collection permissions for a role",
        "operationId": "set_collection_permission",
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "description": "Collection name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetCollectionPermissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Collection permission set successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CollectionPermission"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions - Admin only",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Collection not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/permissions/roles": {
      "get": {
        "tags": [
          "Permissions"
        ],
        "summary": "List all roles",
        "operationId": "list_roles",
        "responses": {
          "200": {
            "description": "Roles retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Vec_Role"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions - Admin only",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Permissions"
        ],
        "summary": "Create a new role",
        "operationId": "create_role",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Role created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Role"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions - Admin only",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/permissions/roles/{role_name}": {
      "get": {
        "tags": [
          "Permissions"
        ],
        "summary": "Get role by name",
        "operationId": "get_role",
        "parameters": [
          {
            "name": "role_name",
            "in": "path",
            "description": "Role name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Role"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions - Admin only",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Role not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_CollectionPermission": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "id",
              "collection_id",
              "role_id",
              "can_create",
              "can_read",
              "can_update",
              "can_delete",
              "can_list",
              "created_at",
              "updated_at"
            ],
            "properties": {
              "can_create": {
                "type": "boolean"
              },
              "can_delete": {
                "type": "boolean"
              },
              "can_list": {
                "type": "boolean"
              },
              "can_read": {
                "type": "boolean"
              },
              "can_update": {
                "type": "boolean"
              },
              "collection_id": {
                "type": "integer",
                "format": "int32"
              },
              "created_at": {
                "type": "string",
                "format": "date-time"
              },
              "id": {
                "type": "integer",
                "format": "int32"
              },
              "role_id": {
                "type": "integer",
                "format": "int32"
              },
              "updated_at": {
                "type": "string",
                "format": "date-time"
              }
            }
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "success": {
            "type": "boolean"
          }
        }
      },
      "ApiResponse_CollectionResponse": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "id",
              "name",
              "schema",
              "is_system",
              "created_at",
              "updated_at"
            ],
            "properties": {
              "created_at": {
                "type": "string",
                "example": "2024-01-01 12:00:00"
              },
              "description": {
                "type": [
                  "string",
                  "null"
                ],
                "example": "Collection for storing product data"
              },
              "display_name": {
                "type": [
                  "string",
                  "null"
                ],
                "example": "Product Collection"
              },
              "id": {
                "type": "integer",
                "format": "int32",
                "example": 1
              },
              "is_system": {
                "type": "boolean",
                "example": false
              },
              "name": {
                "type": "string",
                "example": "products"
              },
              "schema": {
                "$ref": "#/components/schemas/CollectionSchema"
              },
              "updated_at": {
                "type": "string",
                "example": "2024-01-01 12:00:00"
              }
            }
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "ApiResponse_Value": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "data": {

          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "CreateRoleRequest": {
        "type": "object",
        "required": [
          "name",
          "priority"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "priority": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "details": {
            "type": [
              "string",
              "null"
            ]
          },
          "error": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "SetCollectionPermissionRequest": {
        "type": "object",
        "required": [
          "role_id",
          "can_create",
          "can_read",
          "can_update",
          "can_delete",
          "can_list"
        ],
        "properties": {
          "can_create": {
            "type": "boolean"
          },
          "can_delete": {
            "type": "boolean"
          },
          "can_list": {
            "type": "boolean"
          },
          "can_read": {
            "type": "boolean"
          },
          "can_update": {
            "type": "boolean"
          },
          "role_id": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}